# swagger_api_documentation.yaml
openapi: 3.1.0
info:
  title: NirogPath API
  version: 1.0.0
  description: |
    REST API for the **NirogPath Healthcare & OPD Prescription Hub Platform**.
    A comprehensive healthcare management system supporting Patients, Doctors, Clinic Receptionists, and Superuser Admins.
    
    ## Authentication & Security
    Protected endpoints require a HTTP Bearer JWT token passed in the `Authorization` header:
    `Authorization: Bearer <jwt_token>`
    
    Tokens are obtained by authenticating via `POST /api/auth/login` or `POST /api/auth/register`.
    
    ## System Roles & Access Control Matrix
    - **patient**: Browse doctors/hospitals, book consultation slots, pay token deposit via wallet, track live OPD queue position, view e-prescriptions, log daily medication doses, manage family profiles, and ask prescription follow-up questions.
    - **doctor**: Access doctor queue, advance consultation state (`call-next`, `complete`), broadcast running late warnings, set location & availability schedules, write e-prescriptions, reply to patient prescription queries.
    - **receptionist**: Clinic-wide aggregated metrics across doctors, hospital map uploads, walk-in token creation, and patient transfer management.
    - **admin**: Superuser operations — register doctors, create partner hospitals, manage users, and view hospital analytics.
    
    ## User Story Mapping
    | User Story ID | Role | Description | API Endpoints |
    |---|---|---|---|
...
    
...
paths:
...
  /api/waitlist:
    post:
      summary: Join Waitlist
      operationId: join_waitlist_api_waitlist_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WaitlistIn"
        required: true
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:

        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /api/waitlist/{waitlist_id}:
    delete:
      summary: Leave Waitlist
      operationId: leave_waitlist_api_waitlist__waitlist_id__delete
      parameters:
        - name: waitlist_id
          in: path
          required: true
          schema:
            type: string
            title: Waitlist Id
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:

        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
  /api/notifications/{notification_id}/mark-read:
    post:
      summary: Mark Notification Read
      operationId: mark_notification_read_api_notifications__notification_id__mark_read_post
      parameters:
        - name: notification_id
          in: path
          required: true
          schema:
            type: string
            title: Notification Id
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:

        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /api/internal/notify-waitlist:
    post:
      summary: Internal Notify Waitlist
      description: No auth — internal use only. Call from server.py's cancel flow.
      operationId: internal_notify_waitlist_api_internal_notify_waitlist_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotifyWaitlistIn"
        required: true
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:

        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /api/reception/walkin:
    post:
      summary: Add Walkin
      operationId: add_walkin_api_reception_walkin_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalkinIn"
        required: true
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:

        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
  /api/reception/offer-transfer:
    post:
      summary: Offer Transfer
      operationId: offer_transfer_api_reception_offer_transfer_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferTransferIn"
        required: true
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:

        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /api/bookings/{booking_id}/accept-transfer:
    post:
      summary: Accept Transfer
      operationId: accept_transfer_api_bookings__booking_id__accept_transfer_post
      parameters:
        - name: booking_id
          in: path
          required: true
          schema:
            type: string
            title: Booking Id
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:

        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /api/bookings/{booking_id}/decline-transfer:
    post:
      summary: Decline Transfer
      operationId: decline_transfer_api_bookings__booking_id__decline_transfer_post
      parameters:
        - name: booking_id
          in: path
          required: true
          schema:
            type: string
            title: Booking Id
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:

        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    NotifyWaitlistIn:
      properties:
        doctor_id:
          type: string
          title: Doctor Id
        date:
          type: string
          title: Date
        slot_time:
          type: string
          title: Slot Time
      type: object
      required:
        - doctor_id
        - date
        - slot_time
      title: NotifyWaitlistIn
    OfferTransferIn:
      properties:
        booking_id:
          type: string
          title: Booking Id
        target_doctor_id:
          type: string
          title: Target Doctor Id
      type: object
      required:
        - booking_id
        - target_doctor_id
      title: OfferTransferIn
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
    WaitlistIn:
      properties:
        doctor_id:
          type: string
          title: Doctor Id
        date:
          type: string
          title: Date
        slot_time:
          type: string
          title: Slot Time
      type: object
      required:
        - doctor_id
        - date
        - slot_time
      title: WaitlistIn
    WalkinIn:
      properties:
        patient_name:
          type: string
          title: Patient Name
        phone:
          type: string
          title: Phone
        doctor_id:
          type: string
          title: Doctor Id
        urgency:
          type: string
          title: Urgency
          default: routine
      type: object
      required:
        - patient_name
        - phone
        - doctor_id
      title: WalkinIn